file-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pdfkit==0.6.1
python-docx==0.8.11
Werkzeug==3.0.1
